backend/requirements.txt
amqp==2.5.1
anyjson==0.3.3
awscli==1.16.262
beautifulsoup4==4.8.0
billiard==3.6.1.0
boto3==1.9.252
botocore==1.12.252
celery==4.3.0
certifi==2019.6.16
chardet==3.0.4
colorama==0.3.9
coverage==4.5.4
cssselect==1.0.3
Django==2.2.6
django-admin-sortable2==0.7.3
django-appconf==1.0.3
django-celery-beat==1.5.0
django-classy-tags==0.9.0
django-cors-headers==3.1.1
django-filer==1.5.0
django-formtools==2.1
django-fsm==2.6.1
django-fsm-admin==1.2.4
django-ipware==2.1.0
django-js-asset==1.2.2
django-mptt==0.10.0
django-polymorphic==2.0.3
django-post-office==3.2.1
django-rest-auth==0.9.5
django-sekizai==1.0.0
django-select2==6.3.1
django-ses==0.8.12
django-timezone-field==3.1
django-treebeard==4.3
djangorestframework==3.8.2
docutils==0.14
easy-thumbnails==2.6
feedfinder2==0.0.4
feedparser==5.2.1
fpdf==1.7.2
future==0.17.1
googlemaps==3.1.1
gunicorn==19.9.0
html5lib==1.0.1
idna==2.8
importlib-metadata==0.23
jieba3k==0.35.1
jmespath==0.9.4
jsonfield==2.0.2
kombu==4.6.5
lxml>=3.5.0
more-itertools==7.2.0
multidict==4.5.2
newspaper3k==0.2.8
nltk==3.4.4
pdfkit==0.6.1
Pillow==6.1.0
psycopg2-binary==2.8.3
pyasn1==0.4.6
PyPDF2==1.26.0
python-crontab==2.3.9
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.1
PyYAML==5.1.1
redis==3.3.8
requests==2.22.0
requests-file==1.4.3
rsa==3.4.2
s3transfer==0.2.1
sentry-sdk==0.10.2
six==1.12.0
soupsieve==1.9.2
sqlparse==0.3.0
stripe==2.36.2
tinysegmenter==0.3
tldextract==2.2.1
Unidecode==1.0.23
urllib3==1.25.3
vcrpy==2.0.1
vine==1.3.0
webencodings==0.5.1
wrapt==1.11.2
yarl==1.3.0
zipp==0.6.0
